Refactor game structure and refresh UI
