+++ scripts/main.js
@@ -1,0 +1,5 @@
+import { initializeGame } from "./game/engine.js";
+
+document.addEventListener("DOMContentLoaded", () => {
+  initializeGame();
+});
